Moved transform functionality to GeometryService
